Scala 2.13 migration :
  - stricter compiler options
  - remove deprecated things (Stream and string .sorted/.distinct)
  - remove early return statements and prefer find / forall / exists

+++ src/main/scala/EratosthenesSieve.scala
@@ -15,6 +15,6 @@ object EratosthenesSieve {
     sieve(numbers.filter(x => x % first != 0), primes :+ first)
   }
 
-  def apply(limit: Long): Vector[Long] =
-    sieve((2L until limit).toVector, Vector.empty)
+  def apply(limit: Int): Vector[Long] =
+    sieve((2L until limit.toLong).toVector, Vector.empty)
 }
